+++ examples/__main__.py
@@ -1,0 +1,3 @@
+from examples.cli import main
+
+main()
