+++ examples/data/pathway-docs-small/documents.jsonl
