+++ llm_app/__init__.py
@@ -1,6 +1,3 @@
-from llm_app.main import run
 import llm_app.model_wrappers as model_wrappers
-import llm_app.pathway_pipelines as pathway_pipelines
 
-
-__all__ = ["model_wrappers", "pathway_pipelines", "run"]
+__all__ = ["model_wrappers"]
