--- llm_app/__main__.py
@@ -1,3 +1,0 @@
-from llm_app.cli import main
-
-main()
